Move Lazy trie datastructure to its own file (#1871)
Preparation for further developing CegisUnif
